# proiect-mds/openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}}}}},"/status":{"get":{"summary":"Lamp Status","operationId":"lamp_status_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LampStatus"}}}}}}},"/lamp/toggle":{"post":{"summary":"Lamp Toggle","operationId":"lamp_toggle_lamp_toggle_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}}}}},"/lamp/rgb":{"post":{"summary":"Lamp Rgb","operationId":"lamp_rgb_lamp_rgb_post","parameters":[{"required":true,"schema":{"title":"R","type":"integer"},"name":"r","in":"query"},{"required":true,"schema":{"title":"G","type":"integer"},"name":"g","in":"query"},{"required":true,"schema":{"title":"B","type":"integer"},"name":"b","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPError"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/lamp/reading_mode":{"post":{"summary":"Lamp Reading Mode","operationId":"lamp_reading_mode_lamp_reading_mode_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}}}}},"/lamp/wave":{"post":{"summary":"Lamp Wave","operationId":"lamp_wave_lamp_wave_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}}}}},"/lamp/intensity":{"post":{"summary":"Lamp Intensity","operationId":"lamp_intensity_lamp_intensity_post","parameters":[{"required":true,"schema":{"title":"New Intensity","type":"integer"},"name":"new_intensity","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPError"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPError":{"title":"HTTPError","required":["detail"],"type":"object","properties":{"detail":{"title":"Detail","type":"string"}},"example":{"detail":"HTTPException raised."}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"LampStatus":{"title":"LampStatus","required":["last_intensity","intensity","reading_mode","wave"],"type":"object","properties":{"last_intensity":{"title":"Last Intensity","type":"integer"},"intensity":{"title":"Intensity","type":"integer"},"RGB":{"title":"Rgb","type":"array","items":{"type":"integer"},"default":[0,0,0]},"reading_mode":{"title":"Reading Mode","type":"boolean"},"wave":{"title":"Wave","type":"boolean"}}},"Message":{"title":"Message","required":["message"],"type":"object","properties":{"message":{"title":"Message","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}